refactor: math/ を number_theory・combinatorics・linalg に分割
23 ファイルの catch-all 化していた math/ を、test/yosupo/ の分類
（math / number_theory / enumerative_conbinatorics / matrix）に合わせて分割する。
math/ は誤称ではないため改名はせず、modint 等の基盤型のみ残置する。

- number_theory/ 新設: 素数・篩・素因数分解・CRT・floor sum・min_linear・
  mod 平方根・Stern-Brocot 木
- combinatorics/ 新設: combination・enumeration・bell・partition_function・
  offline_binomial_sum
- linalg/ 新設: matrix・berlekamp_massey・linear_recurrence
- math/ 残置: modint・bigint・fraction・hashint・math.hpp
- README のライブラリ構成表を更新

include は lib/ 相対の完全一致パスで一括置換し lib/test 全体を追随。
全 233 テストの -fsyntax-only を確認。
