+++ lib/combinatorics/enumeration.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <vector>
-#include "math/combination.hpp"
+#include "combinatorics/combination.hpp"
 #include "math/modint.hpp"
 
 template <internal::modint mint = modint998>
