+++ lib/combinatorics/partition_function.hpp
