+++ lib/linalg/berlekamp_massey.hpp
