+++ lib/linalg/linear_recurrence.hpp
