+++ lib/number_theory/min_linear.hpp
