+++ lib/number_theory/stern_brocot_tree.hpp
