+++ test/yosupo/math/floor_sum.test.cpp
@@ -1,5 +1,5 @@
 // competitive-verifier: PROBLEM https://judge.yosupo.jp/problem/sum_of_floor_of_linear
-#include "math/floor_sum.hpp"
+#include "number_theory/floor_sum.hpp"
 #include <iostream>
 
 int main(void) {
