+++ test/yosupo/matrix/determinant_of_matrix.test.cpp
@@ -1,7 +1,7 @@
 // competitive-verifier: PROBLEM https://judge.yosupo.jp/problem/matrix_det
 #include <iostream>
 #include <vector>
-#include "math/matrix.hpp"
+#include "linalg/matrix.hpp"
 #include "math/modint.hpp"
 
 using Mint = modint998;
