+++ test/yosupo/matrix/matrix_product.test.cpp
@@ -1,7 +1,7 @@
 // competitive-verifier: PROBLEM https://judge.yosupo.jp/problem/matrix_product
 #include <iostream>
 #include <vector>
-#include "math/matrix.hpp"
+#include "linalg/matrix.hpp"
 #include "math/modint.hpp"
 
 using Mint = modint998;
