+++ test/yosupo/number_theory/counting_primes.test.cpp
@@ -1,6 +1,6 @@
 // competitive-verifier: PROBLEM https://judge.yosupo.jp/problem/counting_primes
 #include <iostream>
-#include "math/prime_counting.hpp"
+#include "number_theory/prime_counting.hpp"
 
 int main(void) {
     std::int64_t n;
