+++ test/yukicoder/2206.test.cpp
@@ -2,8 +2,8 @@
 #include <iostream>
 #include <utility>
 #include <vector>
+#include "combinatorics/offline_binomial_sum.hpp"
 #include "math/modint.hpp"
-#include "math/offline_binomial_sum.hpp"
 
 using Mint = modint998;
